Polyfill dependencies, fix compile errors

+++ src/objects/moves/transform.cpp
@@ -18,7 +18,7 @@ void Transform::Apply(Board* board, bool sim){
     _fromField = _piece->GetField();
     _toField = board->GetSideField(_piece->isWhite());
     _piece->SetField(_toField, sim);
-    board->AddPiece((uint) (_transformInto->GetIdWithoutColor() + 1 - _piece->isWhite()), _fromField);
+    board->AddPiece((_transformInto->GetIdWithoutColor() + 1 - _piece->isWhite()), _fromField);
 }
 
 void Transform::Undo(Board* board, bool sim){
